refactor: separate files into schema and status
