+++ perception_eval/perception_eval/tool/perception_analyzer2d.py
@@ -26,7 +26,7 @@ import numpy as np
 import pandas as pd
 from perception_eval.common.evaluation_task import EvaluationTask
 from perception_eval.common.object2d import DynamicObject2D
-from perception_eval.common.schema import MatchingStatus
+from perception_eval.common.status import MatchingStatus
 from perception_eval.config import PerceptionEvaluationConfig
 from perception_eval.evaluation import DynamicObjectWithPerceptionResult
 import yaml
