+++ perception_eval/test/perception_fp_validation_lsim.py
@@ -19,7 +19,7 @@ from typing import List
 
 from perception_eval.common import ObjectType
 from perception_eval.common.label import AutowareLabel
-from perception_eval.common.schema import get_scene_rates
+from perception_eval.common.status import get_scene_rates
 from perception_eval.config import PerceptionEvaluationConfig
 from perception_eval.evaluation import get_object_status
 from perception_eval.evaluation import PerceptionFrameResult
